+++ src/js/tests/colors.js
@@ -29,15 +29,18 @@ describe('Named colors', async it => {
 
   NamedColors.map(col =>
     it(`Color "${col.toLowerCase()}" parses as a background color`, t =>
-      t.notThrows( () => { const _foo = sm`machine_name: bob; state a: { background-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
+      t.notThrows( () => {
+        const _foo = sm`machine_name: bob; state a: { background-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
 
   NamedColors.map(col =>
     it(`Color "${col.toLowerCase()}" parses as a text color`, t =>
-      t.notThrows( () => { const _foo = sm`machine_name: bob; state a: { text-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
+      t.notThrows( () => {
+        const _foo = sm`machine_name: bob; state a: { text-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
 
   NamedColors.map(col =>
     it(`Color "${col.toLowerCase()}" parses as a border color`, t =>
-      t.notThrows( () => { const _foo = sm`machine_name: bob; state a: { border-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
+      t.notThrows( () => {
+        const _foo = sm`machine_name: bob; state a: { border-color: ${col.toLowerCase()}; }; a->b;`; }) ) );
 
 });